refactored reviews.js, fixed reviewsummaries bug
